Rename package to predic
