+++ predic/DIC_NR_images.py
